Add glace_bin directory to rootwrap exec_dirs
When the cinder backend for glance is enabled and the volume backend
type for cinder is Fibre Channel, glance-api invokes privsep-helper to
execute the commands to access the FC HBA.

Since privsep-helper resides in the glance venv bin directory, this
commit adds that directory to the list of exec_dirs in the rootwrap
config.

Also note that the glance-api service must be installed directly on the
host (i.e. outside of glance_container) in this case because it
attempts to access the host FC HBA through /sys/class/fc_host/ and
/sys/class/scsi_host/.

Upstream-Status: Pending
Signed-off-by: Stephanos Ioannidis <root@stephanos.io>