Store image volumes in service project for cinder backend
When using cinder image store backend, glance stores the image volumes
in the project of the current context.

This can cause problems for public images because the project to which
an image belongs may not be the project where the image is used, and
this results in glance being unable to locate the backing image volume.

This commit updates the glance-api configuration template to specify the
`cinder_store_*` configurations which, when specified, makes glance
store the backing image volumes in the specified project (i.e. service
project) and access the cinder volumes using the specified credentials.

Upstream-Status: Pending
Signed-off-by: Stephanos Ioannidis <root@stephanos.io>